+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 
 -   Optimize styles
+-   Add more unit tests
 
 # 1.0.67 (2025-09-21)